==> notebooks/preprocessing/00_locality_data.ipynb <==
# Locality Data Preprocessing

- Prepares locality data (locality = ZIP code based mapping in Switzerland)
- add altitude of city centers

import pandas as pd
import numpy as np

import geopandas as gp
gp.options.io_engine = "pyogrio"

import pyproj
from shapely.ops import transform

import json

import pandas as pd
from pathlib import Path
import geopandas as gp
import re
gp.options.io_engine = "pyogrio"
from shapely.geometry import Point, Polygon
from datetime import datetime
import plotly.express as px
import rasterio
from rasterio.features import shapes
from shapely.geometry import shape as geoshape
from shapely.geometry import Polygon
import numpy as np
from joblib import Parallel, delayed
from scipy.spatial import cKDTree
import netCDF4
import xarray as xr
import pyproj
from shapely.ops import transform

ws_dir = Path("/mnt/wks3/aschneuwl/workspace")

# ZIP Code - BFS Mapping
Each zip code is assigned to a BFS (politische Gemeinde)

zip_municipality_mapping = pd.read_csv("/home/aschneuwl/workspace/data/geo/AMTOVZ_CSV_LV95/AMTOVZ_CSV_LV95.csv", sep=";", names=["city", "plz", "addDigit", "municipality", "bfs", "kt", "e", "n", "lang", "val"])
zip_municipality_mapping = zip_municipality_mapping.tail(-1)
zip_municipality_mapping.plz = zip_municipality_mapping.plz.apply(lambda x: int(x))
zip_municipality_mapping.bfs = zip_municipality_mapping.bfs.apply(lambda x: int(x))

zip_municipality_mapping[zip_municipality_mapping["plz"] == 8833]

zip_city = zip_municipality_mapping.groupby("plz").agg({"city": set}).reset_index()
zip_city["city"] = zip_city.city.apply(lambda x: ", ".join(sorted(list(x))))
zip_city["plz"] = zip_city.plz.astype(str)

kt = zip_municipality_mapping[["plz", "kt"]].dropna().drop_duplicates().drop_duplicates(subset=["plz"])
kt["plz"] = kt.plz.astype(str)

bfs_per_zip = zip_municipality_mapping.groupby("plz").aggregate({"bfs": "nunique"}).reset_index()
bfs_per_zip[bfs_per_zip.bfs > 1].sort_values("plz")

zip_per_bfs = zip_municipality_mapping.groupby("bfs").aggregate({"plz": "nunique"}).reset_index()
zip_per_bfs[zip_per_bfs.plz > 1]

# Municipalities (Politische Gemeinden)

municipalities = gp.read_file("/home/aschneuwl/workspace/data/geo/swissboundaries3d_2024-01_2056_5728/swissBOUNDARIES3D_1_5_TLM_HOHEITSGEBIET.shp")

lv95 = pyproj.CRS('EPSG:2056')
wgs84 = pyproj.CRS('EPSG:4326')
#municipalities.geometry.simplify(0.5)
project = pyproj.Transformer.from_crs(lv95, wgs84, always_xy=True).transform

for i, shape in municipalities.iterrows():
    if shape.geometry is not None:
        shape.geometry = transform(project, shape.geometry)
        municipalities.iloc[i] = shape

municipalities.to_file('swissBOUNDARIES3D_1_5_TLM_HOHEITSGEBIET.shp')

## Mapping ZIP Code to the largest Municipality (BFS)
A zip code may touch the area of multile municipalities. We assign each zip code to the largest municipality.

bfs_surface_mapping = {municipalities.BFS_NUMMER.values[i]: municipalities.GEM_FLAECH.values[i] for i in range(len(municipalities.BFS_NUMMER.values))}

zip_largest_bfs = {}
for zip_code in list(zip_municipality_mapping.plz.unique()):
    bfs_candidates = zip_municipality_mapping[zip_municipality_mapping.plz == zip_code].bfs.values

    area_sizes = []
    for bfs in bfs_candidates:
        area_sizes.append(bfs_surface_mapping[bfs])
    
    zip_largest_bfs[int(zip_code)] = int(bfs_candidates[np.argmax(area_sizes)])

with open("zip_largest_bfs_2024.json", "w") as fp:
    json.dump(zip_largest_bfs, fp)

# Localities
Each locality has a ZIP code (a ZIP code can be assigned to multiple localities)

localities_zip = gp.read_file("/home/aschneuwl/workspace/data/geo/AMTOVZ_SHP_LV95/AMTOVZ_ZIP.shp")
localities_zip = localities_zip.dissolve(by='ZIP4').reset_index()
localities_zip["location_center"] = localities_zip.geometry.centroid

localities_zip

localities_zip["longitude_x"] = localities_zip.location_center.apply(lambda p: p.x)
localities_zip["latitude_y"] = localities_zip.location_center.apply(lambda p: p.y)

lv95 = pyproj.CRS('EPSG:2056')
wgs84 = pyproj.CRS('EPSG:4326')
#municipalities.geometry.simplify(0.5)
project = pyproj.Transformer.from_crs(lv95, wgs84, always_xy=True).transform

for i, shape in localities_zip.iterrows():
    if shape.geometry is not None:
        shape.geometry = transform(project, shape.geometry)
        shape.location_center = transform(project, shape.location_center)
        localities_zip.iloc[i] = shape

localities_zip = localities_zip.set_crs(wgs84, allow_override=True)

## Altitude

def assign_altitude_to_locality_center(altitude_map_fpath: Path,
                                       locality_gdf: gp.GeoDataFrame,
                                       src_crs: str = "EPSG:21781",
                                       target_crs: str = "EPSG:4326") -> gp.GeoDataFrame:
    
    with rasterio.open(altitude_map_fpath) as src:
        raster_data = src.read(1)
        mask = raster_data != src.nodata
    
        shapes_generator = list(shapes(raster_data, mask=mask, transform=src.transform))
        
        def process_shape(geom, value):
            geom = rasterio.warp.transform_geom(src_crs, target_crs, geom, precision=6)
            
            return geoshape(geom), value
            
        output = Parallel(n_jobs=-1)(delayed(process_shape)(geom, value) for geom, value in shapes_generator)
        polygons, values = zip(*output)
    
    gdf_polygons = gp.GeoDataFrame({'altitude': values, 'geometry': polygons}, crs=target_crs)
    gdf_polygons["center"] = None
    
    lv95 = pyproj.CRS('EPSG:2056')
    wgs84 = pyproj.CRS('EPSG:4326')
    project_to_lv95 = pyproj.Transformer.from_crs(wgs84, lv95, always_xy=True).transform
    project_to_wgs84 = pyproj.Transformer.from_crs(lv95, wgs84, always_xy=True).transform

    def center_helper(shape):
        center = None
        if shape.geometry is not None:
            center = transform(project_to_wgs84, transform(project_to_lv95, shape.geometry).centroid)

        return center

    centers = Parallel(n_jobs=-1)(delayed(center_helper)(shape) for _, shape in gdf_polygons.iterrows())
    gdf_polygons["center"] = centers
    
    # Assign each locality center to the corresponding altitude
    height_search_tree = cKDTree(gdf_polygons.center.apply(lambda geom: (geom.x, geom.y)).tolist())
    distances, indices = height_search_tree.query(locality_gdf.location_center.apply(lambda geom: (geom.x, geom.y)).tolist())
    locality_gdf["altitude"] = gdf_polygons.iloc[indices].altitude.values
    
    # Factor 100 for station matching
    locality_gdf["altitude_100"] = gdf_polygons.iloc[indices].altitude.values * 100

    return locality_gdf

altitude_map_fpath = Path("/home/aschneuwl/workspace/data/geo/height/DHM200.asc")
localities_zip = assign_altitude_to_locality_center(altitude_map_fpath, localities_zip)

localities_zip

localities_zip = localities_zip.merge(kt, left_on="ZIP4", right_on="plz").drop("plz", axis=1)
localities_zip = localities_zip.merge(zip_city, left_on="ZIP4", right_on="plz").drop("plz", axis=1)

ws2 = Path("/mnt/wks3/aschneuwl/workspace/data/preprocessed")

localities_zip.to_parquet(ws2 / Path("swiss_localities_with_altitudes.parquet"))
