==> notebooks/preprocessing/01_meteosuisse_gridded_data.ipynb <==
# Converts gridded data to parquet format

import pandas as pd
from pathlib import Path
import geopandas as gp
import re
gp.options.io_engine = "pyogrio"
from shapely.geometry import Point, Polygon
from datetime import datetime
import plotly.express as px
import rasterio
from rasterio.features import shapes
from shapely.geometry import shape as geoshape
from shapely.geometry import Polygon
import numpy as np
from joblib import Parallel, delayed
from scipy.spatial import cKDTree
import netCDF4
import xarray as xr
import pyproj
from shapely.ops import transform
from typing import List

ws_dir = Path("/home/aschneuwl/workspace/")
ws2_dir = Path("/mnt/wks3/aschneuwl/workspace/data/preprocessed/")

def gridded_to_df(fpath: Path) -> pd.DataFrame:
        # Converts gridded data into pandas DataFrame
        df = xr.open_dataset(fpath).to_dataframe().reset_index().dropna()
        #df[""] = df.apply(lambda x: Point(x.lon, x.lat), axis=1)
        return df

def gridded_data_files_to_parquet(dir: str, measure: str, dst_dir: Path) -> pd.DataFrame:
    gridded_data_dir = Path(dir)
    files = gridded_data_dir.glob(f"{measure}_*.nc")
    
    print("Load")
    gridded_dfs = Parallel(n_jobs=-1, verbose=10)(delayed(gridded_to_df)(fpath) for fpath in files)
    gridded_df = pd.concat(gridded_dfs)
    gridded_df["year"] = gridded_df.time.dt.year
    print("Store...")
    gridded_df.to_parquet(dst_dir / Path(f"{measure}.parquet"), partition_cols=["year"])

# Daily Minimum Temperature

gridded_data_files_to_parquet("/home/aschneuwl/workspace/data/meteo/meteosuisse/grid/NEW_CH01_ALL-DATA-v2024/", "TminD", ws2_dir / Path("meteo"))

# Daily Maximum Temperature

gridded_data_files_to_parquet("/home/aschneuwl/workspace/data/meteo/meteosuisse/grid/NEW_CH01_ALL-DATA-v2024/", "TmaxD", ws2_dir / Path("meteo"))

# Daily Average Temperature

gridded_data_files_to_parquet("/home/aschneuwl/workspace/data/meteo/meteosuisse/grid/NEW_CH01_ALL-DATA-v2024/", "TabsD", ws2_dir / Path("meteo"))

# Daily Precipitation

gridded_data_files_to_parquet("/home/aschneuwl/workspace/data/meteo/meteosuisse/grid/NEW_CH01_ALL-DATA-v2024/", "RhiresD", ws2_dir / Path("meteo"))

# Daily Sunshine Duration

gridded_data_files_to_parquet("/home/aschneuwl/workspace/data/meteo/meteosuisse/grid/NEW_CH01_ALL-DATA-v2024/", "SrelD", ws2_dir / Path("meteo"))

# Plotting

## Daily Maximum Temperature

localities_switzerland_fpath = ws_dir / Path("agecon-thesis/notebooks/00_preprocessing/swiss_localities_with_altitudes.parquet")
localities_swiss = gp.read_parquet(localities_switzerland_fpath)

localities_swiss = gp.read_file("/home/aschneuwl/workspace/agecon-thesis/notebooks/AMTOVZ_ZIP.shp")

lv95 = pyproj.CRS('EPSG:2056')
wgs84 = pyproj.CRS('EPSG:4326')

fig = px.choropleth_mapbox(zip_max_temp, geojson=localities_swiss, featureidkey="properties.ZIP_ID",
                           locations='ZIP_ID', 
                           color='TmaxD',
                           mapbox_style="carto-positron",
                           zoom=6.3, 
                           center = {"lat": 46.8, "lon": 8.5},
                           opacity=0.5,
                           width=1000,
                           height=750,
                           hover_data=["TmaxD"],
                           labels={"TmaxD": "Max Daily Temperature", "ZIP4": "Locality"}
                          )
fig.show()
