==> notebooks/preprocessing/06_filter_milk_data.ipynb <==
import pandas as pd
from pathlib import Path
from joblib import Parallel, delayed
import itertools
import numpy as np
from tqdm import tqdm
from typing import List

workspace = Path("/mnt/wks3/aschneuwl/workspace")

milk_data_fpath = workspace / Path("data/preprocessed/dairy/k33.parquet")
farm_meta_fpath = workspace / Path("data/preprocessed/dairy/swiss_farms_anonymzed.parquet")
calvings_fpath = workspace / Path("data/preprocessed/dairy/k11.parquet")

samples_df = pd.read_parquet(milk_data_fpath)
farm_meta_df = pd.read_parquet(farm_meta_fpath)
calvings_df = pd.read_parquet(calvings_fpath)

samples_df.sampleMethod.value_counts()

sorted(list(samples_df.columns))

years = list(samples_df.year.unique())
breeds =  list(samples_df.animalBreedCode.unique())

years

breeds

samples_df[samples_df.animalBreedCode == "XX"]

def iqa(df: pd.DataFrame, column: str):
    q1 = df[column].quantile(0.25)
    q3 = df[column].quantile(0.75)
    iqr = q3 - q1
    lower_bound = q1 - 1.5 * iqr
    upper_bound = q3 + 1.5 * iqr

    return lower_bound, upper_bound

def iqa_filter(year: int, breed: str, columns: List[str], fpath: Path, min_samples_per_year: int = 1000):
    df = pd.read_parquet(fpath, filters=[[("year", "==", year), ("animalBreedCode", "==", breed)]], engine="pyarrow")
    df = df.dropna(subset=columns, how='any')
    
    filters = []
    for column in columns:
        lower_bound, upper_bound = iqa(df, column)
    
        if (not np.isnan(lower_bound)) and (not np.isnan(upper_bound)):
            filter_str = f"((year == {year}) & (animalBreedCode == '{breed}') & ({column} > {lower_bound}) & ({column} < {upper_bound}))"
            filters.append(filter_str)

    if len(filters) != 0:
        df = df.query(" & ".join(filters))
    else:
        df = None
    
    if (df is not None) and (df.shape[0] < min_samples_per_year):
        df = None

    return df

def filter_yield_for_variable(years: List[int],
                              breeds: List[str],
                              columns: List[str],
                              samples_fpath: Path,
                              farm_meta_df: pd.DataFrame,
                              calvings_df: pd.DataFrame,
                              secondary: List[str]) -> pd.DataFrame:
    """
    Function to filter out outliers from the data (parallelized) . above and below whisker
    """

    # Apply the IQA filter for each breed and year
    combs = list(itertools.product(years, breeds))
    filtered = Parallel(n_jobs=-1)(delayed(iqa_filter)(y,b,columns, samples_fpath) for (y,b) in tqdm(combs))
    filtered = [s for s in filtered if s is not None]

    # base columns
    df_cols = ["animalBreedCode", "animalId", "calvingDate", "farmIdLocationSample", "lactationNumber", "year", "milkingMethod", "sampleWeighingDate", "sampleMethod", "weighingType"]
    
    # primary performance variables
    df_cols.extend(columns)
    
    # secondary performance variables (no IQA applied or na check)
    df_cols.extend(secondary)
    
    df_filtered = pd.concat(filtered)[df_cols]

    # create DIM variable
    df_filtered["days_in_milk"] = df_filtered["sampleWeighingDate"] - df_filtered["calvingDate"]

    #df_filtered = df_filtered.loc[df_filtered["days_in_milk"].dt.days >= 0, :]

    df_filtered = pd.merge(df_filtered, farm_meta_df[["zip", "altitude", "locationType", "farmId"]], left_on='farmIdLocationSample', right_on="farmId", how='left')

    # drop all the samples which were not assigned to a farm
    df_filtered = df_filtered[df_filtered["zip"].notna()].drop(["farmId"], axis=1)

    df_filtered = df_filtered.dropna(subset=["lactationNumber", "year", "milkingMethod", "sampleWeighingDate", "days_in_milk"])

    #df_filtered = pd.merge(df_filtered, calvings_df[["calfId", "calvingDate"]], left_on='animalId', right_on="calfId", how='left')
    #df_filtered["dateOfBirth"] = df_filtered["calvingDate_y"]
    #df_filtered.drop("calvingDate_y", axis=1, inplace=True)
    #df_filtered["calvingDate"] = df_filtered["calvingDate_x"]
    #df_filtered.drop("calvingDate_x", axis=1, inplace=True)
    #df_filtered["age"] = df_filtered["sampleWeighingDate"] - df_filtered["dateOfBirth"]

    #df_filtered["farmId"] = df_filtered["farmId_y"]

    #df_filtered.drop("farmId_y", axis=1, inplace=True)

    #df_filtered.drop("calfId", axis=1, inplace=True)

    df_filtered.to_parquet(workspace / Path(f"/data/preprocessed/dairy/k33_{"_".join(columns)}_iqa_filtered.parquet"), partition_cols=["year"])

    return df_filtered

# Milk

samples_df.milk.isna().sum()

samples_df.fat.isna().sum()

samples_df.protein.isna().sum()

samples_df.somaticCellCount.isna().sum()

samples_df.milkUreaNitrogen.isna().sum()

samples_df.bhbConcentration.isna().sum()

samples_df.aceton.isna().sum()

samples_df.aceton.isna().sum()

samples_df.acetonIr.isna().sum()

samples_df.lactose.isna().sum()

filter_yield_for_variable(years, breeds, primary_dairy_performance_vars, fpath, farm_meta_df, calvings_df)

# Primary Performance Variables

primary_dairy_performance_vars = ["milk", "fat", "protein"]
secondary = ["milkUreaNitrogen", "bhbConcentration", "aceton", "acetonMmol", "acetonIr", "somaticCellCount", "lactose"]

perf_vars =  filter_yield_for_variable(years, breeds, primary_dairy_performance_vars, milk_data_fpath, farm_meta_df, calvings_df, secondary)

perf_vars.shape
