==> notebooks/preprocessing/07_merge_milk_and_weather_data.ipynb <==
import pandas as pd
from pathlib import Path
from joblib import Parallel, delayed
import itertools
import numpy as np
from tqdm import tqdm
from typing import List

workspace = Path("/mnt/wks3/aschneuwl/workspace")

milk_data_fpath = workspace / Path("data/preprocessed/dairy/k33_milk_fat_protein_iqa_filtered.parquet")
weather_fpath = workspace / Path("data/preprocessed/meteo/daily_weather_data_zip_centers.parquet")

w = pd.read_parquet(weather_fpath)

w

years = sorted(list(pd.read_parquet(weather_fpath).year.unique()))

def merge_milk_weather_data(year: int, milk_fpath: Path, weather_fpath: Path):
    milk = pd.read_parquet(milk_fpath, filters=[[("year", "==", year)]], engine="pyarrow")
    weather = pd.read_parquet(weather_fpath, filters=[[("year", "==", year)]], engine="pyarrow")
    df_filtered = pd.merge(milk, weather, left_on=["zip", "sampleWeighingDate"], right_on=["ZIP4", "time"], how='inner')
    df_filtered["year"] = df_filtered["year_x"]
    df_filtered = df_filtered.drop(["time", "ZIP4", "year_x", "year_y"], axis=1)

    return df_filtered

filtered = Parallel(n_jobs=-1)(delayed(merge_milk_weather_data)(y,milk_data_fpath,weather_fpath) for y in tqdm(years))
filtered = [s for s in filtered if s is not None]

filtered = pd.concat(filtered)

filtered.shape

filtered.sampleWeighingDate.min()

filtered.sampleWeighingDate.max()

sorted(filtered.columns)

workspace / Path(f"data/preprocessed/dairy/{milk_data_fpath.name.split(".")[0]}_with_weather.parquet")

filtered.to_parquet(workspace / Path(f"data/preprocessed/dairy/{milk_data_fpath.name.split(".")[0]}_with_weather.parquet"))
